requirements.txt
ogb==1.3.2
torch-scatter==2.0.8
torch-sparse==0.6.7
torch-geometric==2.0.1
mkdocs-material
mkdocstrings
